test/e2e: fold nvidiasmi_xml.go into nvidiasmi_query.go
The split between the two files no longer tracked anything: SMIProcess sat
apart from DiffNoProcessesXML, and ProcessesXML had become a delegation to
GPUReadings.Processes with no production caller left after the scenarios were
converted. Drop it and keep one home for the document assertions.

The encoder/FBC check now returns []string problems like its neighbours, so a
stubbed run reports every wrong reading instead of only the first, and its
element bodies are readings like the rest of the schema — which lets one
diffIntReading helper serve both the #636 and #637 checks and retires
nvidiaSMIInteger.
