+++ tests/e2e/go/assertions/nvidiasmi_query_test.go
@@ -4,6 +4,7 @@
 package assertions
 
 import (
+	"strconv"
 	"strings"
 	"testing"
 
@@ -325,4 +326,166 @@ func TestGPUReadings_ProcessesDecodesConfiguredEntries(t *testing.T) {
 	empty, err := other.Processes()
 	require.NoError(t, err)
 	assert.Empty(t, empty)
+}
+
+// The defect checks below are driven by hand-built documents rather than the
+// captured fixtures: the readings under test are N/A in every capture taken
+// before the NVML getters existed, so the passing case has to be constructed.
+
+// smiXML wraps <gpu> elements in the envelope nvidia-smi emits.
+func smiXML(gpus ...string) string {
+	return `<?xml version="1.0" ?>
+<!DOCTYPE nvidia_smi_log SYSTEM "nvsmi_device_v13.dtd">
+<nvidia_smi_log>
+	<driver_version>550.163.01</driver_version>
+	<attached_gpus>` + strconv.Itoa(len(gpus)) + `</attached_gpus>` +
+		strings.Join(gpus, "") + `
+</nvidia_smi_log>`
+}
+
+// utilizationGPU renders one <gpu> element whose utilization block carries the
+// given jpeg_util and ofa_util bodies verbatim, so a test can inject "35 %",
+// "N/A" or nothing at all.
+func utilizationGPU(id, jpeg, ofa string) string {
+	return `
+	<gpu id="` + id + `">
+		<product_name>NVIDIA A100-SXM4-40GB</product_name>
+		<utilization>
+			<gpu_util>40 %</gpu_util>
+			<memory_util>19 %</memory_util>
+			<encoder_util>0 %</encoder_util>
+			<decoder_util>0 %</decoder_util>` + jpeg + ofa + `
+		</utilization>
+	</gpu>`
+}
+
+const (
+	jpeg35 = "\n\t\t\t<jpeg_util>35 %</jpeg_util>"
+	ofa12  = "\n\t\t\t<ofa_util>12 %</ofa_util>"
+	jpegNA = "\n\t\t\t<jpeg_util>N/A</jpeg_util>"
+	ofaNA  = "\n\t\t\t<ofa_util>N/A</ofa_util>"
+)
+
+func TestDiffJpgOfaUtilizationXML_AcceptsConfiguredPercentages(t *testing.T) {
+	out := smiXML(
+		utilizationGPU("0000:07:00.0", jpeg35, ofa12),
+		utilizationGPU("0000:0F:00.0", jpeg35, ofa12),
+	)
+	require.Empty(t, DiffJpgOfaUtilizationXML(out, 35, 12))
+}
+
+func TestDiffJpgOfaUtilizationXML_RejectsNotAvailableReadings(t *testing.T) {
+	out := smiXML(utilizationGPU("0000:07:00.0", jpegNA, ofaNA))
+
+	problems := DiffJpgOfaUtilizationXML(out, 35, 12)
+	require.Len(t, problems, 2, "both readings are N/A")
+	joined := strings.Join(problems, "\n")
+	assert.Contains(t, joined, `jpeg_util = "N/A"`)
+	assert.Contains(t, joined, `ofa_util = "N/A"`)
+}
+
+// A zeroed default reading must not satisfy a non-zero expectation, and the two
+// values must not be interchangeable — the reason the fixture uses 35 and 12.
+func TestDiffJpgOfaUtilizationXML_RejectsWrongPercentages(t *testing.T) {
+	out := smiXML(utilizationGPU("0000:07:00.0", jpeg35, ofa12))
+
+	assert.Len(t, DiffJpgOfaUtilizationXML(out, 0, 0), 2,
+		"35 %% / 12 %% must not satisfy a zeroed expectation")
+	assert.Len(t, DiffJpgOfaUtilizationXML(out, 12, 35), 2,
+		"transposed expectations must not pass")
+}
+
+// Absent elements decode as empty strings, which must be reported rather than
+// read as 0 %.
+func TestDiffJpgOfaUtilizationXML_ReportsMissingElements(t *testing.T) {
+	out := smiXML(utilizationGPU("0000:07:00.0", "", ""))
+
+	problems := DiffJpgOfaUtilizationXML(out, 35, 12)
+	require.Len(t, problems, 2)
+	joined := strings.Join(problems, "\n")
+	assert.Contains(t, joined, `jpeg_util = ""`)
+	assert.Contains(t, joined, `ofa_util = ""`)
+}
+
+// A getter that answers for only the first device must fail, and the report must
+// name the GPU so the failure points at a device.
+func TestDiffJpgOfaUtilizationXML_ChecksEveryGPU(t *testing.T) {
+	out := smiXML(
+		utilizationGPU("0000:07:00.0", jpeg35, ofa12),
+		utilizationGPU("0000:0F:00.0", jpegNA, ofa12),
+	)
+
+	problems := DiffJpgOfaUtilizationXML(out, 35, 12)
+	require.Len(t, problems, 1, "only the second GPU's jpeg_util is N/A")
+	assert.Contains(t, problems[0], "0000:0F:00.0")
+}
+
+// nvidia-smi can die mid-document and still exit with a partial tree on stdout;
+// truncated XML must be reported as a parse failure, not silently pass for want
+// of any <gpu> elements to check.
+func TestDiffJpgOfaUtilizationXML_RejectsTruncatedOutput(t *testing.T) {
+	full := smiXML(utilizationGPU("0000:07:00.0", jpeg35, ofa12))
+	truncated := full[:strings.Index(full, "<utilization>")]
+
+	problems := DiffJpgOfaUtilizationXML(truncated, 35, 12)
+	require.Len(t, problems, 1)
+	assert.Contains(t, problems[0], "parse nvidia-smi XML")
+}
+
+func TestDiffJpgOfaUtilizationXML_RejectsOutputWithoutGPUs(t *testing.T) {
+	problems := DiffJpgOfaUtilizationXML(smiXML(), 35, 12)
+	require.Len(t, problems, 1)
+	assert.Contains(t, problems[0], "no GPUs")
+}
+
+// statsGPU renders one <gpu> whose encoder_stats, fbc_stats and
+// accounting_mode_buffer_size bodies are the given string, so a test can pass
+// the configured numbers or the N/A the stubs used to return.
+func statsGPU(sessionCount, fps, latency, buffer string) string {
+	return `
+	<gpu id="0000:07:00.0">
+		<accounting_mode_buffer_size>` + buffer + `</accounting_mode_buffer_size>
+		<encoder_stats>
+			<session_count>` + sessionCount + `</session_count>
+			<average_fps>` + fps + `</average_fps>
+			<average_latency>` + latency + `</average_latency>
+		</encoder_stats>
+		<fbc_stats>
+			<session_count>` + sessionCount + `</session_count>
+			<average_fps>` + fps + `</average_fps>
+			<average_latency>` + latency + `</average_latency>
+		</fbc_stats>
+	</gpu>`
+}
+
+func TestDiffEncoderFBCXML_AcceptsConfiguredStats(t *testing.T) {
+	want := EncoderFBCStats{SessionCount: 2, AverageFPS: 30, AverageLatencyUS: 1500}
+
+	problems := DiffEncoderFBCXML(smiXML(statsGPU("2", "30", "1500 us", "4000")), want, want, 4000)
+	assert.Empty(t, problems, strings.Join(problems, "; "))
+}
+
+// The stubbed exports rendered every one of these as N/A, which must be
+// reported as a missing getter rather than compared as a number.
+func TestDiffEncoderFBCXML_RejectsStubValues(t *testing.T) {
+	want := EncoderFBCStats{SessionCount: 2, AverageFPS: 30, AverageLatencyUS: 1500}
+
+	problems := DiffEncoderFBCXML(smiXML(statsGPU("N/A", "N/A", "N/A", "N/A")), want, want, 4000)
+	require.Len(t, problems, 7, "three readings per stats block plus the accounting buffer")
+	joined := strings.Join(problems, "; ")
+	assert.Contains(t, joined, "encoder_stats session_count")
+	assert.Contains(t, joined, "fbc_stats average_latency")
+	assert.Contains(t, joined, "accounting_mode_buffer_size")
+	assert.Contains(t, joined, "getter is missing or unimplemented")
+}
+
+// Encoder and FBC carry separate expectations, so a getter that serves one
+// block's numbers for both must be caught.
+func TestDiffEncoderFBCXML_ChecksBothBlocksAgainstTheirOwnExpectation(t *testing.T) {
+	encoder := EncoderFBCStats{SessionCount: 2, AverageFPS: 30, AverageLatencyUS: 1500}
+	fbc := EncoderFBCStats{SessionCount: 4, AverageFPS: 60, AverageLatencyUS: 900}
+
+	problems := DiffEncoderFBCXML(smiXML(statsGPU("2", "30", "1500 us", "4000")), encoder, fbc, 4000)
+	require.Len(t, problems, 3, "the fbc_stats block matches the encoder expectation")
+	assert.Contains(t, strings.Join(problems, "; "), "fbc_stats session_count = 2, want 4")
 }
